memora-clients/python/pyproject.toml
[project]
name = "memora-client"
version = "0.0.1"
description = "Python client for Memora - Semantic memory system with personality-driven thinking"
authors = [
    {name = "Memora Team"}
]
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "httpx>=0.23.0,<0.29.0",
    "attrs>=22.2.0",
    "python-dateutil>=2.8.0,<3",
]

[tool.uv.build-backend]
module-name = "agent_memory_api_client"
module-root = ""

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
